The user is now debited after purchasing a drink
Cleaned up the how the user table was updated. Made sure I was using the
correct drink price and not an old testing variable. Added a work around
for going to the index after purchasing a drink (as the old ajax method
is no longer working) and close the modal after 5 seconds.